Extracting mock_loader to its own helper file.
